New multimedia plugin jPlayer (jplayer.org) with our own custom skin "zenphotolight" as I found the original ones not fitting. IMPORTANT NOTE: The player look is determined by  pure html/CSS skins. While that is actually a benefit, there are and may be display issues in certain themes if their parent styles overpower the skin ones.
This sadly applies even to the standard skins provided by the player ifself. The zenphotolight custom skin seems to work in all standard themes but only partly in Stopdesign (playlists and audio layout is wrong). This CANNOT be fixed in the skin css itself for all themes. So for certain themes there may be the need to create a custom adjusted skin (or just use flowplayer as before if too lazy). (The css additions uploaded with this commit just center the player).

See also the plugin file main comments if testing.

(BTW, although there exists a 3rd party plugin zenjPlayer this is not a rewrite. The 3rd party plugin was just a simple plain mp3 playlist one lacking many standard plugin features).