Require confirmation of security issues
Setup now has the facility to require the user to acknowledge any
configuration issues that might impact his site security.

This is used now for REGISTER_GLOBALS and for DISPLAY_ERRORS PHP
settings. Thes will show as errors until the user acknowledges that he
is willing to live with the security weaknesses.