limit the number of images cached in parallel
cacheManager now has an option controlling the number of cache resizes
it will allow to procede in parallel. Note that this option does not
control the number of image request processes spawned by the WEB
browser. Rather it serializes the execution of those processes so that
only the defined number will actually execute the i.php script at a
time.