revert pull #48
The update is not yet ready for inclusion:

The flock and mysql lock handling should be separated so that only that
code that is needed will be loaded. (image processing is memory
intensive, so every little bit helps.)

The change relies on mysqli wich may not be present. Indeed the database
handler model needs to be updated with whatever is needed to make this
work so that non-MySQL based implementations might have a chance of
working. At anyrate it needs to use library independent functions.

We would prefer that the lock be implemented as an object to reduce the
function polution of the environment.

The options for processing time limit probably are not something to
expose to our typical user.