Update of portable URLs
GitHub does seem to have issues with branching if the branch gets stale.

This is a re-issue of the changes for portable URLs. This update does
make extensive changes to avoid using the base get and set methods and
use the getxxx, setxxx methods instead.

At some point we may make the base get and set methods private as there
really should be item specific methods, and obviously, the base get and
set methods bypass any object manipulation of the raw data which is
certainy a bad thing.

But for now there are a few places that do not have the appropriate
getxxx setxxx methods, so those get/set calls stay.