reverts recent changes
Note: the database table structure for categories and pages is liniar
while the object structure is hierarchical. This means that it is at
best impractical to use SQL queries to retrieve "published" items
because the item's parent may not be published.

So, for instance, for categories it is needed to have the master
category list include ALL categories to enable the fetch code to test to
be sure that a category does not reside in an unpublished parent (which
implies that the item is not published even if it is marked published.)

The original code did do this processing and any revision MUST take the
hierarchy into consideration.