Use auto text instead of eol=lf (keep LF in the repo)
This is somewhat counter-intiuitive, but it makes some sense. Windows editors will sometimes safe CRLF in files, but we don't want them committed that way, and we want to ignore line ending changes for the most part. If we let git handle this automatically, it will always convert to LF on commit, and keep the repo consistent. Hopefully this will get rid of our problems.