fix #132
This limits "free form" EXIF fields to 8191 characters. Seems a
reasonable limit 'cause what is Zenphoto going to do with bigger fields.
But there appears no specification limit to the size, so if this is too
small....

But probably you will run out of memory if you have many of these
fields....