JPlayer playlist handling dynamic albums
If I understand everything correctly, this is what is needed to fix the
problem.

However, the playlist does not play anything for me. Not for dynamic
albums, not for normal albums, and not even with the 1.4.3 release and
normal albums.

Note also that there is a secondary problem in the original code as the
getFullImageURL() will return a pointer to the alternative file (eg Ogg,
Avi, wmf, etc. I fixed this but the play still does not happen.