fix #175
Note, this just fixes the notification that the release is a debug
build. The list of files is expected. Setup has always checked the file
timestamps of the installation for consistency. If a file's timestamp
varies by "too great" a value it is reported as suspect.

If you run directly from a GigHub repository your file timestamps will
represent when they were last committed in the repository so will likely
vary considerably.

Unfortunately there seems no better check for the validity of the files.
I did try once to make checksums, CRCs, etc. for the files.
Unfortunately at least with PHP those checksums were not consistent
across computer platforms.