Issue #220: leave static image URL cleartext
This is in a function definition in javascript. I can only assume that
these strings are not html decoded by the browser. Anyway, i.php gets
passed the correct strings if cleartext and gets passed the encoded html
entities if the string is encoded.

This will fix the error shown in the debug log. But even with that error
for me all that fails is the display of the image, the video still
plays.