Change text to refer to deleting setup scripts (instead of files)
All that really was a hack threat were the setup PHP scripts. #232
showed us that deleting the CSS was a bad idea, so now only the script
files are deleted. For the 1.4.4.x release we do not change translated
text unless necessary, so that text was left saying setup files (which
is still technically correct)