site closure plugin improvements
The plugin will now work without mod_rewrite.
In addition, if a Zenphoto RSS request is noted it wil redirect to an
RSS feed which says that the feed is suspended.

NOTE: two important things to deal with:
1. The mod_rewrite rules will be updated so that the query parameters
are passed to the redirect script.
2. The user MUST either remove or manually update the closed.php script
in his plugins/site_upgrade folder as the script has been modified such
that there is now a script and a user modifiable closed.htm file that
the script will load for a "normal" gallery request