1.4.5 code added
Again GitHub has made going forward impossible. Maybe it is just the
Windows client, but if you get things even a little out of sync it seems
impossible to get them back.

There really needs to be a command to the effect: "I really know that
this branch head goes with that branch head, believe me"