change Imagick blacklist to a whitelist
This also fixed an issue where Imagick::queryFormats(); was used. Seems
this method is not static at least in the installation of the Zenphoto
Sandbox. A whitelist was choosen because it is much simpler to decide
which formats we want to support than which ones we do not
want--specially since the latter might grow over time as Imagick decides
to support more formats.

Note: In testing on the sandbox site there were problems with creating
images. I do not think this change can have any effect on the caching.
It is limited to the initialization. But maybe some other functions that
were statically used have become dynamic. Haven't been able to view the
error log yet to see what it shows.