Definable rewrite names
This is the first pass at making the "tags" we use for our rewrite URLs
changeable by our users.

Please see the config file for the current list of definitions. These
definitions **should** be used in link generation and are used in the
rewrite rules. I have attempted to find all the places links with the
terms were generated and change them to use the defines.

But most likely I have missed some. Please aid in the testing of this
feature. Change the definitions and see what breaks.