User groups plugin supports custom data filters(fix #287)
This extends the use of admin custom data filters to groups and user
templates. Note that the filter will be invoked now where it was not
before. Probably this is harmless, but filters should probably test to
see if they are called for a user if they do not want to handle groups.

The quota manager now supports groups and templates.