insure that the count of conbinews items represents the number of items that will be got
NOTE:

This simply uses the result of getCombiNews() for the count, caching the
request for efficiency. It makes NO ATTEMPT to insure that the proper
RIGHTS management is applied in getting the items.

It also necessarily assumes that the calls on the count function are
compatible with the calls on the get function, but there is no
justification for this since there are no parameters to countCombiNews
other than the "published" value.