fix #461
The original code was needed for PHP earlier than PHP 5 in order to
insure that the value returned was an object reference. The PHP 5 "new"
operator now returns only a reference, so with PHP 5 the code is not
needed and in fact is deprecated. Since Zenphoto requires PHP 5 the cond
is now corrected to be PHP 5 compliant.