Rework of #505
It is now apparent that page number validation cannot be done without
some participation from themes, at least themes that do "odd-ball"
things (such as having zenpage items on the index page.)

This change adds the ability for themes to participate in the decision
as to if page numbers are valid. Only default and perhaps zpmobile are
simple enough to not require participation.