for now the URL page validation is a Debug Install option
There needs to be much more testing of this feature before it is
unleashed on the world. So it is now active only for themes that opt-in
and when the installation is a Debug Build. (Use the Mark release plugin
to set the build to debug.)

Also, Stopdesign, as a soon to be deprecated theme will not have support
added. (e.g. Stopdesign opts out!)