Cachemanger tab display incorrect
The Cachemanager tab ```Cache images``` display is messed up. Digging
around I noticed that the link for all subtabs in the admin interface
have the css ```id``` set to the tab name. In the case of the ```Cache
images``` tab, the tab name of ```images``` being used as the css
```id``` was picking up some css that was not meant for it. I dug around
some more, but could not find a reason for the css ```id``` being set on
the tab links. So, I removed it. All the other subtabs appear to still
be working correctly. However before implementing this pull request, you
may wish to do a more thorough review.