#731
Option names are now case insensitive. The SQL for checking the database
for them has always been case insensitive. But the array that cached
them used the name verbatum for the index. This meant that for coding
mistakes where the option name was the wrong case you could get a bad
value stored that was not overridden.

The options array indexes are all now lower cased and the value used to
index it is lowercased before checking if the option is in memory.