Revert comment form changes from 281291aa9857e765c8dd5f0fe3886d537cc6f535
The address fields in comments have absolutely nothing to do with
whether there there are user address fields. Do not be confused by the
sharing of code (now not done) and the sharing of function.

I suppose that the comment form could extend the comment database and
use fields, but there seemed no need. Instead it stores the address info
in the custom_data field. But since the comment object is unique to
comment handling there is no conflict in doing so. There is not any
filter that allows other plugins to use the comment object custom_data
field.