#791
Note that this is a good example of the needed changes. There are three
places where the instance of the translated phrase with a quote was
used. In one of them it is not actually possible to simply change the
single quote in the js function call to a double one because the js text
itself is enclosed in double quotes.