the issues cited in 8fc2a492160ca1cdaf1f8dccccffaba079176042 are more ubiquitous
The mainline admin pages all force a page reload after handling
"actions" insuring that the state of the gallery is clean and
up-to-date. Unfortunately the zenpage equivalents do not do this,
instead flowing directly into the page rendering.

This is wrong, but untangling the processing from the rendering is
complicated by the coding practices of the implementer.

So far as I can see, only the cms object is potentially "unclean" at the
start of the page rendering, so the Quick-and-dirty fix is to
re-instantiate it before the rendering begins.

But note that there may be other structures that might now or in the
future be unclean, so this is a potential for bugs.