Fixed the logic of detecting login success; fix zp_handle_password
The bug was caused just by the use of || instead of &&, but in addition, we should never be passing the password through any sort of sanitization since the only thing done with it is hashing it to oblivion. Sanitization could incorrectly dispose of special characters that should be valid in a password.