Rationalize use of the "show" column
The computations for future publishing and for expiration were both
inconsistent among objects and often obscurely conceived.

This change insures that the "show" column reflects the current
published state.

Note that there is no longer the [ill-conceived] concept of a
page/article that is published, but not really since the publish date is
in the future. Things are published or they are not.