getNotViewableImages uses a wrong clause when determining which images should be hidden
getNotViewableImages uses a wrong clause when determining which images
should be hidden. It checks for show = 0 and albumid = #. This
doesn't exclude the images which have show =1 but are contained in
albums the user doesn't have access to. This makes the printAllTagsAs(
with checkAccess = true) print the wrong count as it will count all
items with the tag but not exclude the items the user doesn't have
access to.