reorganize the authorization class
Makes lib-auth a class base for user management. The standard install
will load class-auth which loads lib-auth and extends the _Authority and
_Administrator classes into the required Zenphoto_Authority and
Zenphoto_Administrator classes.

When Zenphoto20 starts up it will look for
"..plugins/alt/class-auth.php" and if found will load it in lieu of the
standard class-auth.php. This allows developers to extend/override the
standard administrative classes for such purposes as single login
implementations or use of other credential authorities for
authentication.