fix #960
Always do the full analysis and don't fix symptoms.

Certainly the existence of a sort key value as index  was not making the
images out of order. Rather the problem came when getting images by
ordinal index.

Solution is for the objects $this->images array to have ordinal indexes
not to remove the [used elsewhere} sort key index from the return value
of the sorting functions.

Since there were multiple ways to get this non-ordinal index  and
multiple places where an ordinal index was assumed the legacy zenphoto
"fix" just masked the issue for it to pop-up later.