Resolves zenphoto/zenphoto#944
That issue is closed but not fixed in zenphoto 1.4.10. I had the same problem, not due to using subdomains for something else than language selection, but due to using an IP address (not domain name) for a development server.
The changes also:
- replace some calls to getOption by a single call + use of a local variable
- correct indentation.
Fix based on ZenPhoto20/ZenPhoto20@929a662f0d214bb687fa79371f3d677136c2fdfd.
Tested with zenpage theme.