prevent erroeous XSRF block
Some scripts do not re-launch after saves, etc. so are not operating
from the same clean slate each time. This can cause misceleanous
administrator properties to be different between the generation of the
XSRF token and the check of it. This change insures that irelevant
properties are ignored when creating the token.