fix #1027
Note: while automataically protecting the scripts is possible it is not
straight forward as you need to be careful under what circumstances you
do this. Specially in a multi-user/multi-threaded environment you can
pull the rug out from someone.

Anyway, it is a but that the scripts get restored without the explicit
action of an administrator. Zenphoto20 has fixed that but, so the
automatic handling is not needed.