Adds optional extra security validation
XSRF defending can now have two parts to the token. This allows for
adding a part of the URL to the token generation effectively preventing
forged URLs

Note that only administrators could successfully forge URLs in the first
place, so this maybe over kill. Use it when altering the URL might
expose the software to injection hacks.