.htaccess for backup folder
Note also that 9cd0359d669cef19c43bdf282a4b2a9dd06fc1c4 did not merge
the change to load zenphoto.js only for logged in users. The reason is
quite simple--there are javascript functions in this file that are used
under all circumstances.

It is really a pitty that the current support team for legacy zenphoto
does not understand the software they are trying to support.