PREV/NEXTNews links now relative to catagory
If a news article page has been reached by clicking a link on an article
link on a news category page the category context will be preserved.
Thus the PREV and NEXT links will be to artiicles within that category.

fix #1087
This was a perfectly reasonable request. Categories are just as much
nesting/organizational elements as albums.