some proper use of functions
truncate_string() simply chops off the end of the string. It should not
be used for text with HTML tags. It also makes no sense to
html_encodeTagged() such a string.

shortenContent() is used when there is html content that should be
preserved (if in the shortened part.) So it makes no sense to use it on
"bare" content nor to strip tags from the string once shortened.

Titles generally should not have html tags, so should be using "bare"
strings and truncate_string().

Content output should normally respect the html in the source so should
be using html_encodeTagged(shortenContent(....))