adds option to set TinyMCE "entity_encode" option
This is a "fix" for #1094. It is not really a fix since the problem is
actually an issue with MySQL current versions. But this does let one set
the "entity_encode" option to "numeric"  and use numeric strings like "
&#127774;" for these Emoji.

Note that if you are not using TinyMCE then the " &#127774;" encoding
would already work.