warn about secure cookies
If one has been using HTTPS on the back end and sets the server protocol
to HTTP there may be browser cookeis around that are https only. If so,
the users will not be able to log in using HTTP until those cookies are
cleared. But the flag to switch to HTTPS will be cleared, so the
back-end references will not automatically switch to HTTPS