relax unique images requirements
The unique images option no longer requires a mod_rewrite suffix. URLs
to images without either a mod_reqrite_suffix or a proper image suffix
will follow the path of unique images and try to find the image. If it
is found and 'unique images' is not set the browser will be permanantly
redirected to the link with the image suffix.

#731