change load sequence vs locale setting
The locale cannot really be set until auth.php has run, otherwise thre is no user of record and his settings cannot be honored.

Also, plugins that use gettext before the locale is set will get it wrong 99% of the time. So we now have a delayed actions array to accomodate these strings.