fix cookie issue
If GALLERY_SESSIONS is not set, sessions are not available on the front-end. So any cookie also stored in the SESSION variable may be out of sync with reality.

If the cookie itself is removed, it still may be ghosted in the SESSION variable.

So, don't look at the SESSION variable for a cookie unless GALLERY_SESSIONS is set.