prep for organization change
The software will now detect if the repository owner has changed to netPhotoGraphics.

Through use of defines and similar methodology the proper "owner" will be placed in URLs that reference gitHub.

Note that this change needs to be revisited sometime after the repository change in order to make the owner name a constant rather than querying GitHub.