Fix unique_option constraint being removed
This is a problem that continues to happen to me and others sometimes. There are other mentions of it on here and the zenphoto forum, including from me in 2014 (and it happened to me twice this year)
e.g.
https://forum.zenphoto.org/discussion/906506/options-table-spam
https://forum.zenphoto.org/discussion/1410812/millions-of-records-in-the-options-table-duplicate-options-continuously-created
https://github.com/zenphoto/zenphoto/issues/532

I finally looked into the code and I believe I have found the cause of it - on a high-traffic environment, if the database update script is run while users are still browsing the website, there could be a race-condition:

1) Drop the constraint (line 2333)
2) A duplicate entry is added to the table
3) Try to add the constraint (line 2334) but it fails because of the duplicate

This commit has two attempted fixes that should make it less likely to happen:
1) The existing drop and alter command are run as one query now which should reduce or eliminate the chance that another db operation happens in between them. Some databases may even guarantee the safety of this operation with their internal locks but I don't know if that can be relied upon in a high-traffic environment.
2) Attempt to run the command again with `IGNORE` which, when it is supported* will make the constraint resolve its duplicates automatically.

A more bulletproof fix would likely require locking zenphoto while these scripts run, but I think this will be an improvement if not a fix for some users. It does fix it for me.

*The caveat is that the `ALTER IGNORE` syntax was removed in MySQL 5.7.4, so this is only a fix for users of earlier versions, which AFAIK is a lot of people because of Linux distro defaults.